refactor(api): Optimizar las incidencias y la ruta
