+++ index.js
@@ -7,7 +7,7 @@ const { mongoConnection } = require('./DB')
 dotenv.config();
 const app = express();
 const Ruta = require('./Endpoints/Ruta')
-
+const Incidencia = require('./Endpoints/Incidencia')
 const Authentication = require('./Endpoints/Authentication')
 
 mongoConnection(process.env.MONGODB_URI)
@@ -31,6 +31,8 @@ app.use('/docs', swaggerUi.serve, swaggerUi.setup(specs))
 
 app.use('/auth', Authentication)
 app.use('/ruta', Ruta)
+app.use('/incidencia', Incidencia)
+
 
 /**
  * @openapi
